Added some very basic support for an integration test
I didn't want to write a whole bunch of unit tests on the compiler output as the chance of the output changing every now and then is pretty high.
What I do want to test with every change is that the generated code still properly decodes/encodes data.

In order to test that we actually have to generate and build some cpp code that uses the generated code. So that's exactly what I did. We generate some code, a cmake file, run cmake and run the executable.
We then take the return value of that and consider the the test passed if the return value is 0. Not super sophisticated, but good enough for now.