Changed generator to add _view to the generated view class and removed copy and assignment ctors.
Copy and assignment constructors will lead to trouble as this class is
only a view on existing data. The data in the class doesn't really exist
and is just there to give the methods of the class a pointer to the raw
data.

Copying an object would make a new object that only contains the first
byte of that raw data.