#20 #21 #22 A lot of work in this commit
Should have done intermediate commits, but I didn't.
Implemented support for non-standard type integers throughout. May need to add some extra testing, but so far so good.
We did end up with code that's a little messy so the next thing should probably be to clean up a little bit.