#25 Refactored cpp generator code
Added classes for the different field types to separate the code generation code a bit. It's not that important now, but as we'll add support for more field types it should help keep the code more readable.
Not completely done yet; I think some more code can be moved from the BuilderGenerator into the type class(es)