#39 cleaning up integration tests
Pulled all of the communal code into a base class and added a test suite so I can easily run all of the integration tests.
I had to fumble around a bit to get this to work in a way where I could easily run the individual unit tests and still have them be recognized as such by my IDE while still being able to run all of the tests at once. That's why I ended up with multiple inheritance.
I know that python unittest has its own test discovery, but the requirement that all modules should be top level importable made this too much of a hassle to use. Hence the test suite.
The down side is that we manually need to add the tests to the suite.