#41 Added preliminary support for enum to model
It's a little hackier than I'd like it to be, but I think I'm going to need to implement nested types and imports to really get my head around how this should work. I'm going to first try to get the enumeration working and see if I can make it all a little nicer later.