#68 WIP Have Evaluate return smallest expression
Evaluate used to return a number or None. Now working on having it return the smallest possible expression. If that's a number, it will be a number literal. If we can't evaluate to a number because there is a dynamic field (e.g. a reference) in there, it will try to resolve any parts of the expression it can and create the smallest possible expression (e.g. some_field / (5 + 3) will resolve into some_field / 8)

Still need to update the equality and ternary operators and probably clean up a little