#73 Added unary minus operator
Because recognizing negative values in the lexer prevents us from capturing the binary minus operator. The model does not include the unary minus operator yet. I simply hacked that right into the parse tree walker, but we'll have to properly add it to the model to do semantic analysis on the expression.
We also still have to add static analysis rules for the enumerators because grammatically any expression can be used as a value for the enumerator. Which is obviously not what we want.