#72 Removed support for the power operator
Adding support for the power operator was quite tricky. Having a signed exponent for example can lead to a floating point result even of both operands are integers. While thinking about this, I didn't really see any real use cases for the power operator so it seemed easier to just drop it altogether. If I'm mistaken and there are use cases, we can always add it later.